+++ node-repository/src/test/java/com/yahoo/vespa/hosted/provision/restapi/responses/maintenance.json
@@ -43,9 +43,6 @@
       "name": "OsUpgradeActivator"
     },
     {
-      "name": "ParkedExpirer"
-    },
-    {
       "name": "PeriodicApplicationMaintainer"
     },
     {
